+++ public/results.php
@@ -266,7 +266,7 @@ if (count($results[$p]) > 0) {
                           <span class="caret"></span></button>
                           <ul class="dropdown-menu">
                               <li class="small"><a href="advanced.php?index=<?php echo $esIndex; ?>&amp;index2=<?php echo $esIndex2; ?>&amp;submitted=true&amp;p=1&amp;path_parent=<?php echo rawurlencode($file['path_parent']); ?>"><i class="glyphicon glyphicon-filter"></i> filter (non-recursive)</a></li>
-                              <li class="small"><a href="simple.php?index=<?php echo $esIndex; ?>&amp;index2=<?php echo $esIndex2; ?>&amp;submitted=true&amp;p=1&amp;q=path_parent:<?php echo rawurlencode(escape_chars($file['path_parent'] . '*')); ?>"><i class="glyphicon glyphicon-filter"></i> filter (recursive)</a></li>
+                              <li class="small"><a href="simple.php?index=<?php echo $esIndex; ?>&amp;index2=<?php echo $esIndex2; ?>&amp;submitted=true&amp;p=1&amp;q=path_parent:<?php echo rawurlencode(escape_chars($file['path_parent'])); ?> OR path_parent:<?php echo rawurlencode(escape_chars($file['path_parent']) . '\/*'); ?>"><i class="glyphicon glyphicon-filter"></i> filter (recursive)</a></li>
                               </ul>
                       </div>
               <!-- end path buttons -->
@@ -420,7 +420,7 @@ else {
   <div class="row">
     <div class="alert alert-dismissible alert-info col-xs-8">
       <button type="button" class="close" data-dismiss="alert">&times;</button>
-      <i class="glyphicon glyphicon-exclamation-sign"></i> <strong>Sorry, no items found :(</strong> Change a few things up and try searching again. <a href="#" onclick="window.history.go(-1); return false;">Go back</a>.
+      <i class="glyphicon glyphicon-exclamation-sign"></i> <strong>Sorry, no items found.</strong> Change a few things up and try searching again. <a href="#" onclick="window.history.go(-1); return false;">Go back</a>.
     </div>
   </div>
 </div>