+++ src/diskover/Diskover.php
@@ -11,7 +11,7 @@ use Elasticsearch\ClientBuilder;
 error_reporting(E_ALL ^ E_NOTICE);
 
 // diskover-web version
-$VERSION = '1.5.0-beta.8';
+$VERSION = '1.5.0-beta.10';
 
 
 function connectES() {